+++ src/Thycotic.RabbitMq.Helper.PSCommands/Installation/GetRabbitMqInstallerCommand.cs
@@ -2,6 +2,7 @@
 using System.Management.Automation;
 using System.Threading;
 using Thycotic.RabbitMq.Helper.Logic;
+using Thycotic.RabbitMq.Helper.Logic.IO;
 
 namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
 {
@@ -50,6 +51,12 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
              ValueFromPipeline = true,
              ValueFromPipelineByPropertyName = true,
              ParameterSetName = ParameterSets.Offline)]
+        [Parameter(
+            Position = 0,
+            Mandatory = true,
+            ValueFromPipeline = true,
+            ValueFromPipelineByPropertyName = true,
+            ParameterSetName = ParameterSets.Prepare)]
         [Alias("OfflinePath")]
         public string OfflineRabbitMqInstallerPath { get; set; }
 
@@ -64,6 +71,10 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
              ValueFromPipeline = true,
              ValueFromPipelineByPropertyName = true,
              ParameterSetName = ParameterSets.Online)]
+        [Parameter(
+            ValueFromPipeline = true,
+            ValueFromPipelineByPropertyName = true,
+            ParameterSetName = ParameterSets.Prepare)]
         [Alias("ForceDownload")]
         public SwitchParameter Force { get; set; }
 
@@ -79,15 +90,25 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
              ValueFromPipeline = true,
              ValueFromPipelineByPropertyName = true,
              ParameterSetName = ParameterSets.Online)]
+        [Parameter(
+            ValueFromPipeline = true,
+            ValueFromPipelineByPropertyName = true,
+            ParameterSetName = ParameterSets.Prepare)]
         [Alias("Mirror")]
         public SwitchParameter UseThycoticMirror { get; set; }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="GetRabbitMqInstallerCommand"/> class.
+        /// Gets or sets whether to prepare for offline install.
         /// </summary>
-        public GetRabbitMqInstallerCommand()
-        {
-        }
+        /// <value>
+        ///   <c>true</c> if [force download]; otherwise, <c>false</c>.
+        /// </value>
+        /// <para type="description">Gets or sets whether to prepare for offline install.</para>
+        [Parameter(
+            ValueFromPipeline = true,
+            ValueFromPipelineByPropertyName = true,
+            ParameterSetName = ParameterSets.Prepare)]
+        public SwitchParameter PrepareForOfflineInstall { get; set; }
 
         /// <summary>
         ///     Processes the record.
@@ -97,22 +118,56 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
         {
             if (!string.IsNullOrWhiteSpace(OfflineRabbitMqInstallerPath))
             {
-                WriteVerbose(string.Format("Using offline installer path {0}", OfflineRabbitMqInstallerPath));
-
-                if (!File.Exists(OfflineRabbitMqInstallerPath))
+                if (PrepareForOfflineInstall)
                 {
-                    throw new FileNotFoundException("Installer does not exist");
-                }
+                    WriteDebug($"Preparing offline installer path {OfflineRabbitMqInstallerPath}");
 
-                if (PrerequisiteDownloader.CalculateMD5(OfflineRabbitMqInstallerPath) != InstallationConstants.RabbitMq.InstallerChecksum)
+                    if (Force)
+                        WriteVerbose("Forcing download");
+
+                    WriteVerbose("Downloading Erlang");
+
+                    var downloader = new PrerequisiteDownloader();
+
+                    var downloadUrl = UseThycoticMirror
+                        ? InstallationConstants.Erlang.ThycoticMirrorDownloadUrl
+                        : InstallationConstants.Erlang.DownloadUrl;
+
+                    downloader.Download(CancellationToken.None, downloadUrl,
+                        OfflineRabbitMqInstallerPath, InstallationConstants.Erlang.InstallerChecksum, Force, 5,
+                        WriteDebug, WriteVerbose, (s, exception) =>
+                        {
+                            throw exception;
+                        },
+                        progress =>
+                        {
+                            WriteProgress(new ProgressRecord(1, "Erlang download in progress", "Downloading")
+                            {
+                                PercentComplete = progress.ProgressPercentage
+                            });
+                        });
+
+                }
+                else
                 {
-                    throw new FileNotFoundException("Installer checksum does not match");
+                    WriteVerbose(string.Format("Using offline installer path {0}", OfflineRabbitMqInstallerPath));
+
+                    if (!File.Exists(OfflineRabbitMqInstallerPath))
+                    {
+                        throw new FileNotFoundException("Installer does not exist");
+                    }
+
+                    if (PrerequisiteDownloader.CalculateMD5(OfflineRabbitMqInstallerPath) !=
+                        InstallationConstants.RabbitMq.InstallerChecksum)
+                    {
+                        throw new FileNotFoundException("Installer checksum does not match");
+                    }
+
+                    if (File.Exists(RabbitMqInstallerPath))
+                        File.Delete(RabbitMqInstallerPath);
+
+                    File.Copy(OfflineRabbitMqInstallerPath, RabbitMqInstallerPath);
                 }
-
-                if (File.Exists(RabbitMqInstallerPath))
-                    File.Delete(RabbitMqInstallerPath);
-
-                File.Copy(OfflineRabbitMqInstallerPath, RabbitMqInstallerPath);
             }
             else
             {
@@ -144,6 +199,7 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
         {
             public const string Offline = "Offline";
             public const string Online = "Online";
+            public const string Prepare = "Prepare";
         }
     }
 }
